requirements.txt
pdfplumber
faiss-cpu
numpy
streamlit
openai
langchain 
logging